# Chore: Optimize cache clear in dev mode