<!-----------------------------------------------------------------------------
Thank you for contributing to the PrestaShop project!

Please take the time to edit the "Answers" rows below with the necessary information.

Check out our contribution guidelines to find out how to complete it:
https://devdocs.prestashop-project.org/9/contribute/contribution-guidelines/#pull-requests

For type and category see:
https://devdocs.prestashop-project.org/9/contribute/contribution-guidelines/pull-requests/#type--category
------------------------------------------------------------------------------>

| Questions         | Answers
| ----------------- | -------------------------------------------------------
| Branch?           | 9.0.x
| Description?      | While in dev mode, I noticed a real struggle when I wanted to clear my Symfony cache. After further study, it is mainly on translations / twig templates cache warming. With this little changes, I managed to make it really fast. Before, I couldn't even warm my cache `(Error: Allowed memory size of 536870912 bytes exhausted (tried to allocate 32768 bytes)`. With this change, I no longer have this error and the cache clearing operation is much faster. (See details below)
| Type?             | bug fix / improvement / new feature / refacto
| Category?         | CO
| BC breaks?        | no
| Deprecations?     | no
| How to test?      | Run command `bin/console cache:clear -vvv`
| UI Tests          | Please run UI tests and paste here the link to the run. As we support MySQL and MariaDB in our tests, you have to launch 2 campaigns (by choosing both). [Visit the UI Tests repo and follow instructions](https://github.com/PrestaShop/ga.tests.ui.pr/).
| Fixed issue or discussion?     | ~
| Related PRs       | ~
| Sponsor company   | ~

**Before:**
```bash
time bin/console cache:clear -vvv

 // Clearing the cache for the dev environment with debug true

 // Cache is fresh.

 // Warming up optional cache...

 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\TranslationsCacheWarmer" completed in 1481.04ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\RouterCacheWarmer" completed in 86.45ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\AnnotationsCacheWarmer" completed in 1.99ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\SerializerCacheWarmer" completed in 0.54ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\ValidatorCacheWarmer" completed in 84.39ms.
 [INFO] "Symfony\Bundle\SecurityBundle\CacheWarmer\ExpressionCacheWarmer" completed in 0.71ms.
 [INFO] "Symfony\Bundle\TwigBundle\CacheWarmer\TemplateCacheWarmer" completed in 23270.34ms.
 [INFO] "PrestaShopBundle\Routing\AnonymousRouteProvider" completed in 17.20ms.
 [INFO] "Symfony\UX\TwigComponent\CacheWarmer\TwigComponentCacheWarmer" completed in 5.21ms.

 // Removing old build and cache directory...

 // Finished

 [OK] Cache for the "dev" environment (debug=true) was successfully cleared.
```

> real    0m36.389s
> user    0m14.368s
> sys     0m21.758s

**After:**
```bash
time bin/console cache:clear -vvv

 // Clearing the cache for the dev environment with debug true

 // Cache is fresh.

 // Warming up optional cache...

 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\RouterCacheWarmer" completed in 97.74ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\AnnotationsCacheWarmer" completed in 2.49ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\SerializerCacheWarmer" completed in 0.80ms.
 [INFO] "Symfony\Bundle\FrameworkBundle\CacheWarmer\ValidatorCacheWarmer" completed in 33.90ms.
 [INFO] "Symfony\Bundle\SecurityBundle\CacheWarmer\ExpressionCacheWarmer" completed in 0.41ms.
 [INFO] "PrestaShopBundle\Routing\AnonymousRouteProvider" completed in 0.64ms.
 [INFO] "Symfony\UX\TwigComponent\CacheWarmer\TwigComponentCacheWarmer" completed in 1.00ms.

 // Finished

 [OK] Cache for the "dev" environment (debug=true) was successfully cleared.
```

> real    0m4.899s
> user    0m2.200s
> sys     0m2.680s