@ChillCode 2025-09-26
Is not better to use --no-warmup parameter?

Also the issue is described here: https://github.com/PrestaShop/PrestaShop/discussions/38977