@jolelievre 2025-09-28
Hmm, can you not reach the same thing by using the `--no-warmup` command option? Or maybe with the `--no-optional-warmers` that skips even more warmup?