@ChillCode 2025-10-02
It's a normal behavior as axplined on this discussion https://github.com/PrestaShop/PrestaShop/discussions/39653

That command is warming up everything:

If we open a template inside twig cache folder we will see thousands of files with code like this:

>     public function getSourceContext(): Source
>     {
>         return new Source("", "@AdminNewTheme/node_modules/is-number/index.js", "/var/www/prestashop.local/public_html/admin-dev/themes/new-theme/node_modules/is-number/index.js");
>     }

Why we need to lint a js file? No way.

Before using file_name_pattern directive:

<img width="923" height="34" alt="image" src="https://github.com/user-attachments/assets/0a169ff4-4ee6-4a1f-90b0-ad0125270d70" />

After using file_name_pattern directive:

<img width="915" height="30" alt="image" src="https://github.com/user-attachments/assets/3a5cc637-9755-4a57-ac1d-1aede82b7015" />

The time is in accordance the size.