@jolelievre 2025-10-02
Yeah, ideally we should loop through all modules and add a specific twig namespace This would reduce the number of files parsed and warmed up But, the thing is we didin't force this convention when the feature was introduced, so if we change it now it will be a breaking change

We could introduce this on 9.1 though, something that would replace the generic `@Modules/module_name/views/templates/path_to_my_twig_file.html.twig` with multiple small `@Modules.module_name/path_to_my_twig_file.html.twig` for example, and deprecate the former namespace but it can only be removed in the next major 10.0